refactor: give the endpoint its param table back
