--- packages/render-core/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "render-core",
-  "//version": "Never published, but the version must exist: packing a consumer (blobatar-cli) resolves every workspace range, including devDependencies, and an unversioned workspace member fails the pack.",
-  "version": "0.0.0",
-  "private": true,
-  "//private": "The deliberate exception to \"a Package is publishable\" (ADR-0009): a private support package. Never published — wrangler bundles it into the endpoint Worker (apps/api imports it by path, so the standalone clone the Deploy button makes needs no installable name), and the CLI build inlines it into the distributed bin. Its exports map points at TypeScript source because every consumer is a bundler that transpiles.",
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "test": "bun test",
-    "check": "bun test"
-  },
-  "dependencies": {
-    "blobatar": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/bun": "latest"
-  }
-}
